+++ audiolazy/__init__.py
@@ -62,7 +62,6 @@ iterables, lists, generators, list comprehensions and decorators.
 
 This package was created by Danilo J. S. Bellini and is a free software,
 under the terms of the GPLv3.
-
 """
 
 import os
@@ -78,34 +77,47 @@ __modules__ = sorted({_mname.split(".")[0]
                      })
 
 # Imports all modules to the main namespace
-_mdocs_pairs = []
 _pkg_name = os.path.split(__path__[0])[1]
+__modules_refs__ = []
 __all__ = []
 for _mname in __modules__:
   exec "from .{0} import *".format(_mname)
   _mref = sys.modules[".".join([_pkg_name, _mname])] # The module, actually
+  __modules_refs__.append(_mref)
   __all__ += _mref.__all__ # __all__ don't include module names
 
-  # Creates a table with the docstring data from each module
-  _mdoc = [_row.strip() for _row in _mref.__doc__
-                                         .split("\n\n")[0] # First paragraph
-                                         .strip() # No empty lines
-                                         .splitlines() # As a list of strings
-          ]
-  _mdocs_pairs.append((_mname, _mdoc))
+# Updates module docs (now small_doc and rst_table are already in namespace)
+_mdocs_pairs = []
+_maxmlen = 78 - max(len(_mname) for _mname in __modules__)
+for _mname, _mref in zip(__modules__, __modules_refs__):
+  _mref.__doc__ += "\nSummary of module contents: \n\n"
+  _maxclen = 78 - max(len(_obj_name) for _obj_name in _mref.__all__)
+  _table = [(_obj_name, small_doc(getattr(_mref, _obj_name),
+                                  max_width=_maxclen)
+            ) for _obj_name in _mref.__all__]
+  _mref.__doc__ += "\n".join(rst_table(_table, ("Name", "Description")))
+  _mref.__doc__ += "\n"
+
+  # Creates table w/ docstring data from each module (w/ lazy_misc.small_doc)
+  _mdocs_pairs.append((_mname, small_doc(_mref, max_width=_maxmlen)))
 
 # Edits the package docstring (with audiolazy.lazy_misc.rst_table)
+__doc__ += "\nSummary of package modules: \n\n"
 __doc__ += "\n".join(rst_table(_mdocs_pairs, ("Module", "Description")))
+__doc__ += "\n"
 
 # Remove references just for namespace clean-up
 if "_mname" in locals(): # Not all interpreters keep the for loop variable
   del _mname
+if "_mref" in locals():
+  del _mref
 del _mdocs_pairs
 del _pkg_name
-del _mref # At least one module should exist, so this should have existed
-del _mdoc
-if "_row" in locals():
-  del _row
+del _table # At least one module should exist, so this should have existed
+if "_obj_name" in locals():
+  del _obj_name
+del _maxclen
+del _maxmlen
 del os
 del sys